requirements.txt
numpy==2.2.4
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.2.1
PySide6==6.9.0
PyYAML==6.0.2
torch>=2.5.1
ultralytics==8.3.58
flake8==7.1.2
pytest-qt==4.4.0
pytest-timeout==2.3.1
markdown==3.7